updated backend

+++ invoice-tracker-backend/src/main/java/net/talaatharb/invoicetracker/models/Role.java
@@ -11,11 +11,11 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Role {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
 
-    @Enumerated(EnumType.STRING)
-    @Column(length = 20)
-    private ERole name;
+	@Enumerated(EnumType.STRING)
+	@Column(length = 20)
+	private ERole name;
 }
